requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
mccabe==0.7.0
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
